Extract shared versioning core from asset module
